Move poeditor client to pkg to allow inclusion
