+++ internal/rest/v1/project.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/labstack/echo/v4"
 	"github.com/sirupsen/logrus"
-	"github.com/uniwise/parrot/internal/poedit"
+	"github.com/uniwise/parrot/pkg/poedit"
 )
 
 type Content struct {
